requirements.txt
Flask==2.3.2
supabase==1.0.4
python-dotenv==1.0.0
Werkzeug==2.3.6
